DEX-505 lift state and use uppy hook to prevent creation of multiple uppy instances and lose track of assets upon click of back button
